A new crawling function is added
The function "crawl_yahoo" has the same functions as "crawl_company_price", but the former allow us to crawl historical price up to many years without keep sending requests. The form of price_data shall be the same. Please check and test the new crawler (And compare to the old one).